Revert "Editor Deprecation: Consolidate the editor URL selectors (#47344)"
This reverts commit 24e5b0759b78ae6619c2647c4e2627b759b04389.
