+++ client/state/selectors/get-editor-url.js
@@ -1,14 +1,15 @@
 /**
  * Internal dependencies
  */
-import { isEligibleForGutenframe } from 'calypso/state/gutenberg-iframe-eligible/is-eligible-for-gutenframe';
-import { getSiteAdminUrl, getSiteSlug } from 'calypso/state/sites/selectors';
-import { getEditorPath } from 'calypso/state/editor/selectors';
-import { addQueryArgs } from 'calypso/lib/route';
-import isVipSite from 'calypso/state/selectors/is-vip-site';
+import getGutenbergEditorUrl from 'calypso/state/selectors/get-gutenberg-editor-url';
+import isClassicEditorForced from 'calypso/state/selectors/is-classic-editor-forced';
+import isEligibleForGutenframe from 'calypso/state/gutenberg-iframe-eligible/is-eligible-for-gutenframe';
+import { getSiteAdminUrl } from 'calypso/state/sites/selectors';
 
 export const getEditorUrl = ( state, siteId, postId = null, postType = 'post' ) => {
-	if ( ! isEligibleForGutenframe( state, siteId ) ) {
+	// isEligibleForGutenframe deals with the server side checks for classic editor plugin etc.
+	// isClassicEditerForced has the client side checks. We should combine these!
+	if ( ! isEligibleForGutenframe( state, siteId ) || isClassicEditorForced( state, siteId ) ) {
 		const siteAdminUrl = getSiteAdminUrl( state, siteId );
 		let url = `${ siteAdminUrl }post-new.php?post_type=${ postType }`;
 
@@ -16,23 +17,10 @@ export const getEditorUrl = ( state, siteId, postId = null, postType = 'post' ) 
 			url = `${ siteAdminUrl }post.php?post=${ postId }&action=edit`;
 		}
 
-		if ( ! isVipSite( state, siteId ) ) {
-			url = addQueryArgs( { calypsoify: '1' }, url );
-		}
-
 		return url;
 	}
 
-	if ( postId ) {
-		return getEditorPath( state, siteId, postId, postType );
-	}
-
-	const siteSlug = getSiteSlug( state, siteId );
-
-	if ( 'post' === postType || 'page' === postType ) {
-		return `/${ postType }/${ siteSlug }`;
-	}
-	return `/edit/${ postType }/${ siteSlug }`;
+	return getGutenbergEditorUrl( state, siteId, postId, postType );
 };
 
 export default getEditorUrl;
